Preliminary work on fixing enums.
The join operation for enums should switch from inluscion to top,
when max_elems is reached. Top is represented by Exc ([],r), for some
range r.
To implement this switch, a range field was added to Inc as well,
similiar to the patch for DefExc.

The range should now describe the range of the integer data type of the
variable rather than the most precise range for the current abstract
value.

Work needs to be done to return values with the correct range in
mult and div, limit the length of lists returned by of_interval,
and to check the widening.
Test case 35-enums.c still fails when run with option
--set ana.int.enums_max 2.